feat: allow not dropping previous updates when starting polling

+++ aiogram_bot_template/bot.py
@@ -165,7 +165,8 @@ async def aiogram_on_shutdown_webhook(dispatcher: Dispatcher, bot: Bot) -> None:
 
 
 async def aiogram_on_startup_polling(dispatcher: Dispatcher, bot: Bot) -> None:
-    await bot.delete_webhook(drop_pending_updates=True)
+    if config.DROP_PREVIOUS_UPDATES:
+        await bot.delete_webhook(drop_pending_updates=True)
     await setup_aiogram(dispatcher)
     dispatcher["aiogram_logger"].info("Started polling")
